Enable parallel computation
This patch extracts many of the global state objects constructed when
creating new DFAs. This allows creation of multiple DFAs in parallel
for multithreaded applications.

In addition, abort functions have been replaced where possible with
error code propagation to prevent process crashing when an error
is encountered on a single thread